splits schemas for separate requests.py and responses.py files
